Refactor flink app history info method to new Controller and Service.
