+++ streamx-console/streamx-console-service/src/main/java/com/streamxhub/streamx/console/core/controller/ApplicationController.java
@@ -42,7 +42,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.IOException;
 import java.io.Serializable;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -232,14 +231,6 @@ public class ApplicationController {
         StorageType storageType = Application.getStorageType(executionMode);
         boolean upload = applicationService.upload(file, storageType);
         return RestResponse.create().data(upload);
-    }
-
-
-    @PostMapping("listUploads")
-    @RequiresPermissions("app:create")
-    public RestResponse listUploadJars() {
-        List<String> jars = applicationService.listUploadJars(StorageType.LFS);
-        return RestResponse.create().data(jars);
     }
 
 }
