+++ streamx-console/streamx-console-webapp/src/api/application.js
@@ -112,7 +112,3 @@ export function startLog (params) {
 export function checkJar(params) {
   return http.post(api.Application.CHECKJAR, params)
 }
-
-export function listUploads() {
-  return http.post(api.Application.LISTUPLOADS)
-}
